Create result sub-module; shift storage+result related functionality around
