+++ pypesto/optimize/util.py
@@ -1,7 +1,5 @@
 """Utility functions for :py:func:`pypesto.optimize.minimize`."""
-import datetime
-import os
-import binascii
+
 from pathlib import Path
 from typing import List
 import h5py
@@ -9,8 +7,6 @@ import h5py
 from ..engine import Engine, SingleCoreEngine
 from ..objective import HistoryOptions
 from ..store.save_to_hdf5 import get_or_create_group
-from ..store import write_result
-from ..result import Result
 from .optimizer import OptimizerResult
 
 
@@ -108,39 +104,3 @@ def postprocess_hdf5_history(
 
     # reset storage file (undo preprocessing changes)
     history_options.storage_file = storage_file
-
-
-def autosave(filename: str,
-             result: Result,
-             store_type: str,
-             overwrite: bool = False):
-    """
-    Save the result of optimization, profiling or sampling automatically.
-
-    Parameters
-    ----------
-    filename:
-        Either the filename to save to or "Auto", in which case it will
-        automatically generate a file named
-        `year_month_day_{type}_result.hdf5`.
-    result:
-        The result to be saved.
-    store_type:
-        Either `optimize`, `sample` or `profile`. Depending on the
-        method the function is called in.
-    overwrite:
-        Whether to overwrite the currently existing results.
-    """
-    if filename is None:
-        return
-
-    if filename == "Auto":
-        time = datetime.datetime.now().strftime("%Y_%d_%m_%H_%M_%S")
-        filename = time+f"_{store_type}_result_" \
-                        f"{binascii.b2a_hex(os.urandom(8)).decode()}.h5"
-    # set the type to True and pass it on to write_result
-    to_save = {store_type: True}
-    write_result(result=result,
-                 overwrite=overwrite,
-                 filename=filename,
-                 **to_save)
