+++ pypesto/predict/__init__.py
@@ -13,3 +13,4 @@ from .result import (
     PredictionConditionResult,
 )
 from .task import PredictorTask
+from . import constants
