+++ pypesto/profile/__init__.py
@@ -9,8 +9,6 @@ from .approximate import (
     approximate_parameter_profile)
 from .options import (
     ProfileOptions)
-from .result import (
-    ProfilerResult)
 from .validation_intervals import (
     validation_profile_significance)
 from .util import (
