+++ pypesto/profile/profile.py
@@ -5,7 +5,7 @@ from ..engine import Engine, SingleCoreEngine
 from ..optimize import Optimizer
 from ..problem import Problem
 from ..result import Result
-from ..optimize.util import autosave
+from ..store import autosave
 from .profile_next_guess import next_guess
 from .options import ProfileOptions
 from .util import initialize_profile
@@ -25,7 +25,7 @@ def parameter_profile(
         next_guess_method: Union[Callable, str] = 'adaptive_step_regression',
         profile_options: ProfileOptions = None,
         progress_bar: bool = True,
-        filename: Union[str, None] = "Auto"
+        filename: Union[str, None] = "Auto",
 ) -> Result:
     """
     Call to do parameter profiling.
