+++ pypesto/result/__init__.py
@@ -1,0 +1,13 @@
+"""
+Result
+======
+
+The pypesto.Result object contains all results generated by
+the pypesto components. It contains sub-results for
+optimization, profiling, sampling.
+"""
+
+from .result import Result
+from .optimize import OptimizerResult, OptimizeResult
+from .profile import ProfilerResult, ProfileResult
+from .sample import McmcPtResult, SampleResult
