+++ pypesto/result/sample.py
@@ -1,3 +1,5 @@
+"""Sampling result."""
+
 import numpy as np
 from typing import Iterable
 
@@ -33,16 +35,18 @@ class McmcPtResult(dict):
     iterations (i.e., the chain length), and `n_par` the number of parameters.
     """
 
-    def __init__(self,
-                 trace_x: np.ndarray,
-                 trace_neglogpost: np.ndarray,
-                 trace_neglogprior: np.ndarray,
-                 betas: Iterable[float],
-                 burn_in: int = None,
-                 time: float = 0.,
-                 auto_correlation: float = None,
-                 effective_sample_size: float = None,
-                 message: str = None):
+    def __init__(
+        self,
+        trace_x: np.ndarray,
+        trace_neglogpost: np.ndarray,
+        trace_neglogprior: np.ndarray,
+        betas: Iterable[float],
+        burn_in: int = None,
+        time: float = 0.,
+        auto_correlation: float = None,
+        effective_sample_size: float = None,
+        message: str = None,
+    ):
         super().__init__()
 
         self.trace_x = trace_x
@@ -88,3 +92,10 @@ class McmcPtResult(dict):
 
     __setattr__ = dict.__setitem__
     __delattr__ = dict.__delitem__
+
+
+class SampleResult:
+    """Result of the sample() function."""
+
+    def __init__(self):
+        pass
