+++ pypesto/sample/__init__.py
@@ -25,7 +25,6 @@ from .adaptive_parallel_tempering import (
 )
 from .pymc3 import Pymc3Sampler
 from .emcee import EmceeSampler
-from .result import McmcPtResult
 from .diagnostics import (
     auto_correlation,
     effective_sample_size,
