+++ pypesto/sample/pymc3.py
@@ -6,7 +6,7 @@ import logging
 from ..objective import History
 from ..problem import Problem
 from .sampler import Sampler
-from .result import McmcPtResult
+from ..result import McmcPtResult
 
 logger = logging.getLogger(__name__)
