+++ pypesto/store/save_to_hdf5.py
@@ -1,4 +1,5 @@
 """Include functions for saving various results to hdf5."""
+
 import os
 import logging
 from typing import Union
