+++ test/base/test_ensemble.py
@@ -14,6 +14,7 @@ from pypesto.predict.constants import (AMICI_STATUS,
 from pypesto.ensemble.constants import MEAN, WEIGHTED_SIGMA
 from pypesto.predict import AmiciPredictor
 from pypesto.engine import MultiProcessEngine
+import pypesto.optimize as optimize
 
 
 def test_ensemble_from_optimization():
@@ -30,9 +31,9 @@ def test_ensemble_from_optimization():
 
     problem = pypesto.Problem(objective=objective, lb=lb, ub=ub)
 
-    optimizer = pypesto.optimize.ScipyOptimizer(options={'maxiter': 10})
+    optimizer = optimize.ScipyOptimizer(options={'maxiter': 10})
     history_options = pypesto.HistoryOptions(trace_record=True)
-    result = pypesto.optimize.minimize(
+    result = optimize.minimize(
         problem=problem, optimizer=optimizer,
         n_starts=n_starts, history_options=history_options)
 
@@ -153,8 +154,8 @@ def get_ensemble_prediction(max_size: int = 2,
     """
     problem = create_petab_problem()
 
-    optimizer = pypesto.optimize.ScipyOptimizer()
-    result = pypesto.optimize.minimize(
+    optimizer = optimize.ScipyOptimizer()
+    result = optimize.minimize(
         problem=problem, optimizer=optimizer,
         n_starts=2, filename=None)
