+++ test/profile/test_profile.py
@@ -48,7 +48,7 @@ class ProfilerTest(unittest.TestCase):
             # check result
             self.assertTrue(
                 isinstance(result.profile_result.list[i_run][0],
-                           profile.ProfilerResult))
+                           pypesto.ProfilerResult))
             self.assertEqual(len(result.profile_result.list), i_run+1)
             self.assertEqual(len(result.profile_result.list[i_run]), 2)
 
@@ -126,7 +126,7 @@ class ProfilerTest(unittest.TestCase):
             filename=None)
 
         self.assertIsInstance(result.profile_result.list[0][1],
-                              profile.ProfilerResult)
+                              pypesto.ProfilerResult)
         self.assertIsNone(result.profile_result.list[0][0])
 
         # 2nd run of profiling, appending to an existing list of profiles
@@ -142,7 +142,7 @@ class ProfilerTest(unittest.TestCase):
             filename=None)
 
         self.assertIsInstance(result.profile_result.list[0][0],
-                              profile.ProfilerResult)
+                              pypesto.ProfilerResult)
 
         # 3rd run of profiling, opening a new list, using the default algorithm
         result = profile.parameter_profile(
@@ -155,7 +155,7 @@ class ProfilerTest(unittest.TestCase):
             filename=None)
         # check result
         self.assertIsInstance(result.profile_result.list[1][0],
-                              profile.ProfilerResult)
+                              pypesto.ProfilerResult)
         self.assertIsNone(result.profile_result.list[1][1])
 
     def test_extending_profiles(self):
@@ -182,10 +182,10 @@ class ProfilerTest(unittest.TestCase):
         # check result
         self.assertTrue(
             isinstance(result.profile_result.list[0][0],
-                       profile.ProfilerResult))
+                       pypesto.ProfilerResult))
         self.assertTrue(
             isinstance(result.profile_result.list[0][1],
-                       profile.ProfilerResult))
+                       pypesto.ProfilerResult))
 
     def test_approximate_profiles(self):
         """Test for the approximate profile function."""
@@ -196,7 +196,7 @@ class ProfilerTest(unittest.TestCase):
             n_steps=n_steps)
         profile_list = result.profile_result.list[-1]
         assert profile_list[0] is None
-        assert isinstance(profile_list[1], profile.ProfilerResult)
+        assert isinstance(profile_list[1], pypesto.ProfilerResult)
         assert np.isclose(profile_list[1].ratio_path.max(), 1)
         assert len(profile_list[1].ratio_path) == n_steps
         assert profile_list[1].x_path.shape == (2, n_steps)
